requirements.txt
fastapi>=0.129.0
psycopg2-binary>=2.9.11
pydantic>=2.12.0
python-dotenv>=1.2.0
SQLAlchemy>=2.0.46
pymysql>=1.1.0
openpyxl>=3.1.0
uvicorn>=0.30.0
